Reworked vertical scrolling. Scroll up/down in the middle.
This will actually keep the scroll wherever we trigger scrolling. If we
trigger scrolling on the last item, then we will scroll on up until the
first item is in view. If we started to scroll in the middle, then we
will scroll on up until the first item is in view. This basically just
keeps scrolling up and down consistent.