Always use provided token for remote control requests. Fixes #309
If a remote control provides a (transient) token and connection info, always
use it instead of whatever we have in our own server manager. This is important
for two main reasons. First, the flinging user may have different access rights
than our own user, and so the media may be unavailable to our own user. Second,
the media should be considered watched by the flinging user, not our own.

It's very important that we not remember the connection/token that's provided
beyond the playback of that item/queue. However, we need to make sure that the
token is used for multiple requests. We need to use the provided token to
fetch the original item metadata, then any play queue operations, including
future refreshes of the play queue, and any timeline requests during playback.
This seems tricky, but ends up being fairly straightforward. We can always
create a connection based on the parameters, and then create a server for
that connection, but specifically avoid adding that server to our server
manager. We make the initial (container) key request using that server, which
means that the item will refer to our "temporary" server. The PQ will also
end up with a reference to that server, which means that all items fetched
for the PQ will have that reference. Timeline requests use the server and
connection referenced from the item, so they end up correct as well. As soon
as playback stops, there are no references to the server and it gets garbage
collected.

Note that we do still fall back to using our own server if the controller
doesn't provide a token. This is currently the case for at least Web and
iOS, so we have to handle it.