Add support for multipart video playback. Closes #155
After mentally laboring over how to do this for far too long, here's what
I came up with. I was tempted to make it more seamless by creating multiple
content metadata objects and then playing through them just like we play
through multiple items in a play queue. There were problems with that approach
though. In addition to the fear that some unnoticed code was depending on one
CMD for one Media, things like remote control expect to be able to hit Next
and move to the next media item, not the next part. Trying to special case
that seemed like it was going to be worse than special casing elsewhere.

Still, the check we did in the old channel felt dirty and like something to
be avoided. Now, when we're building CMD objects for a media item, we build
one for every part, and link them to one another. We essentially have a linked
list of CMD objects for the media, one for each part. When we finish playing
a video, we already have some logic that checks to see if there's another video
to be played. That has been tweaked to first check for additional parts linked
off the current object.