Hack up the makefile to make `pkg` work, update README.
The application packager is the bane of my existence. It appears to be
completely broken in firmware 6.2 right now. While trying to make it work, I
noticed that the make target for packaging doesn't work either. There were
some simple things that were simply broken, but I also couldn't make it
reliably work with curl. It turns out to be much simpler using httpie, though
that's an unfortunate dependency. Since I think I'm the only one using this
anyway, I've made that change. We can discuss though.

While I was in there, I tweaked some of the make targets (e.g. `make pass`
instead of `make test` to create a PlexPass package; `make public` instead
of `make rel` for the public channel) and updated the README.