Fix crash when removing playing track from PQ list. Fixes #359
There are a few ways to fix this, but this seems like the easiest along
with fixing similar future issues.

We were running into a race condition with the audio player triggers
and screen updates. We normally update the previous track when we get a
"playing" event from the audio player. In this case, we destroyed the
previous track before the next track started playing. There were a few
options we could have used to fix this, like setting the previous track
invalid (not playing), or set the next track as the current playing
track, before we receive the event, but in the long run, this should fix
other similar issues.