Add support to specify the dropdown spacing.
The spacing between the button and dropdown (flyout) has been based
around the buttons parent spacing. This doesn't work for all
scenarios. There's a chance the button is not part of a container (no
parent). or the possibility we want to space multiple buttons further
apart than we want the dropdown spaced from the button.