Handle filtering more endpoints (hubs and pivots). #81
This is simplified now that we do not need to handle compound
filters. The endpoint itself may contain compound filters, and that is
fine since we'll just override it. The filterBox will use "CUSTOM" as
the filter title, and "SORT" if there is a custom sort. This will
probably need to be tuned.