Add a generic GlyphButton component.
Name still under review. This will add support to create a standard
button with a glyph (second label) right justified of the first
label. Will be used for the 'Clear Filters' button with a delete glyph
on the right. It will probably be used as well for any button with a
checkmark.