Refactor custom buttons in dropdowns.
 * AddCallableButton: Supply it with a button function and arguments. It
   will be created on the fly when the dropdown is expanded.

Initially I had hoped to create a button and push it to the available
dropdown options, but sadly we have memory constraints and are required
to destroy components, which ends up invalidating references we need
when toggling the dropdown (e.g. roFont). Instead we'll employ the
`AddCallableButton` method that creates a callable and pushes it to the
available options, to be created on the fly when needed and without
worry of invalidating required references. So this will basically allow
us to handle adding custom buttons to the dropdown without much change
to the dropdown overlay. The main caveat is if we need to customize the
button, we'll still have to add logic to the dropdown overlay - refer to
SetDropDownPosition in the dropdown overlay.