Refactor dropdown positioning.
Add the option to specifically set dynamic left/right positioning, with
a preferred direction. Previously this was default, and we needed a way
to be able to force the direction specified. It's now required to set
the dynamic positioning instead of just assuming that's what we want.

related b023803085dff58e6e840fa83e41894d177c987c #368