Reset the nextComponentID in a better place (fix refocus).
There are times we refresh a screen and do not call `Init()`, which
resets the `nextComponentID`. In that case, the ID will continue to
increment, and we'll try to match an ID we shouldn't when refocusing. It
should be better to reset the ID while we destroy a screens components,
which is something we always call before we call `GetComponents()`.

Fixes plexinc/roku-client-issues#17