Add focus helper to the album screen (action bar). #418
This is a (working) example that we'll probably want to embrace on all
of our screens, although it's possible we may just add these helpers to
individual screens due to deviation. This will help refocus to the last
button we came from in the action bar, if we are returning to the
left (coming from the item list).