Simplify our VBox scrolling calculations.
Essentially we just need to shift everything to the position of our
scrolling trigger, while accounting for the first and last shift. There
are some new additions coming that require us to shift up or down by an
arbitrary amount, so we cannot assume we'll always scroll up or down by
one component.
 e.g. vertical jump list / manually focusing item within the VBox