Add focus helper to move to the last use filter box button.
This will help the user get back to the last used filter button
quickly after they make a selection and the grid refreshes. e.g. User
selects a sort, the grid refreshes and the focus is now on the first
grid item. The user can now press up once, and will be focused on the
button they made their previous selection from. This should help toggle
sort orders more quickly, but also works for the filters and types
button.