Add `type=13` to all the photo hubs.
I don't see any harm in adding the type to all photo hubs, but this is
specifically done to fix the `photo.recent` hub. It only uses a
sort (not filters), and when a sort is used, the results are viewed as
folders, not photos. If a filter is in use, the results are the same as
if we had already included `type=13`.