Add support for custom filter view type (photos folder view).
This one will probably need UX review/approval.

Override the view `type` if a given url doesn't include one. This is
mainly to workaround oddities with photo libraries / folder view. When
a type type isn't specified for a photo library, the results are
returned by folder view (not an available view type to choose). Once the
user changes the type or filter, it will now be sorted by photos, with
no way to get back to the folder view. Another issue, is that we'll
display "photos" as the type which is obviously a lie.