Revert view type when clearing filters, unless manually set. Fixes #471
The default type when accessing a photo library is `folders` which isn't
available as a default type from the server. We are faking it, by
initially setting the type to `folders`. Once someone filters, it
changes the type to `photos`, but once one clears the filters, the type is
still set to `photos`. We should try and revert back to the initial view
type, as long as the user didn't change the type manually.

Fixes plexinc/roku-client-issues#20